Add support for LATEST version and Maven SNAPSHOT handling.
git-svn-id: https://scm.ops4j.org/repos/ops4j/projects/pax/runner@4899 9b982a3c-3ae5-0310-a4bc-d9a3335569bd

+++ src/main/java/org/ops4j/pax/runner/Downloader.java
@@ -78,8 +78,8 @@ public class Downloader
         }
 
         File localRepo = new File( System.getProperty( "user.home" ), ".m2/repository" );
-
         File localCache = new File( localRepo, path );
+
         for( String repository : m_repositories )
         {
             File md5File = getMD5File( localCache );
@@ -124,6 +124,23 @@ public class Downloader
                 }
             }
         }
+
+        copyFile( localCache, destination );
+    }
+
+    public void copyFileFromLocalRepository( String path, File destination, boolean force )
+        throws IOException
+    {
+        NullArgumentException.validateNotEmpty( path, "path" );
+        NullArgumentException.validateNotNull( destination, "destination" );
+
+        if( destination.exists() && !force )
+        {
+            return;
+        }
+
+        File localRepo = new File( System.getProperty( "user.home" ), ".m2/repository" );
+        File localCache = new File( localRepo, path );
 
         copyFile( localCache, destination );
     }
